Break out the JoystickHistoryView
